Fix simulation of RLWIMIX/RLWINMX
Shift right operation for rotation needs to be done unsigned.

+++ src/ppc/simulator-ppc.cc
@@ -2228,7 +2228,7 @@ void Simulator::InstructionDecode(Instruction* instr) {
       int me = instr->Bits(5, 1);
       int rc = instr->Bit(0);
       // rotate left
-      int result = (rs_val << sh) | (rs_val >> (32-sh));
+      int result = (rs_val << sh) | (((unsigned int)rs_val) >> (32-sh));
       int mask = 0;
       if (mb < me+1) {
         int bit = 0x80000000 >> mb;
@@ -2268,7 +2268,7 @@ void Simulator::InstructionDecode(Instruction* instr) {
       int me = instr->Bits(5, 1);
       int rc = instr->Bit(0);
       // rotate left
-      int result = (rs_val << sh) | (rs_val >> (32-sh));
+      int result = (rs_val << sh) | (((unsigned int)rs_val) >> (32-sh));
       int mask = 0;
       if (mb < me+1) {
         int bit = 0x80000000 >> mb;
